[mod] #12 - MeetingStatus 패키지 이동

+++ app/src/main/java/com/promiseeight/www/domain/model/MeetingMain.kt
@@ -6,11 +6,10 @@ data class MeetingMain (
     val confirmedTime: String?,
     val hostName: String,
     val joinedUserCount: Int,
-    val meetingId: Int,
+    val meetingId: Long,
     val meetingName: String,
     val meetingStatus: String,
-    val minimumAlertMembers: Int,
+    val minimumAlertMembers: Long,
     val votingUserCount: Int,
     val yaksokiType: String
-
 )
